work in progress -- moving out of the AuthContext and into the requireAuth
